==> 02_coarse_classing.ipynb <==
#### Module Summary
> This module does coarse classing of the development data and reduces further features with < 3% Information Value. The cut-off value can be re-configured from the configuration file. \
> the coarse classing output of this module will need manual intervention to group 'similar' bins together for the next module - fine_classing
>
> **Input Files**
> 1. Developemnt and Validation data with reduced features as obtained form 01_feature_reduction module (pickle)
> 2. Feature Readuction Summary (xlsx)
>
> **Output Files**
> 1. Development & Validation data with reduced features (pickle)
> 2. Feature Reduction Summary (xlsx)
> 3. Coarse Classing Data (xlsx)

<h4> 2.1 Import Modules & Parameters

import logit_config as cfg
exec(open('module_imports.py').read())

<h4> 2.2 Import Data

with open(f"{outpath}/model_data.pickle", 'rb') as f:
    model_data_pckl_dict = pickle.load(f)
dev_data = copy.deepcopy(model_data_pckl_dict['dev_data_out1'])
val_data = copy.deepcopy(model_data_pckl_dict['val_data_out1'])
print(f"Development Data Shape: {dev_data.shape}")

<h4> 2.3 Coarse Classing

# %%time

# Get Variable List for which coarse classing needs to be done
var_list = [x for x in dev_data.columns.tolist() if x not in non_pred_varlist]

# Coarse Classing
dev_data_woe = copy.deepcopy(dev_data)
coarse_classing_df = coarse_classing(dev_data, dev_data_woe, resp_var, var_list, special_val_list)

# Update the Character vars Grouping
char_varlist = [x for x in dev_data.select_dtypes(exclude=np.number).columns.tolist() if x not in non_pred_varlist]

if len(char_varlist) > 0:
    man_enc_map_df = create_cat_groups(dev_data, resp_var, char_varlist)
    coarse_classing_df = coarse_classing_df.merge(man_enc_map_df[['feature', 'class', 'grp']], left_on=['VAR_NAME', 'VAR_BINS'], right_on=['feature', 'class'], how='left')
    coarse_classing_df['FINE_BIN_NUM'] = np.where(coarse_classing_df['grp'].isnull(), coarse_classing_df['FINE_BIN_NUM'], coarse_classing_df['grp']).astype(int)
    coarse_classing_df.drop(['feature', 'class', 'grp'], axis=1, inplace=True)
coarse_classing_df.head()

<h4> 2.4 Information Value Check

# Compute IV
iv_df = pd.DataFrame(coarse_classing_df.groupby('VAR_NAME')['INFO_VAL'].sum().rename('IV')).rename_axis('feature').reset_index().sort_values('IV', ascending=False).reset_index(drop=True)

# Drop Variables
# iv_drop_df = iv_df[iv_df['IV'] <= iv_cutoff]
iv_drop_df = iv_df[iv_df.index >= 100]
dev_data.drop(iv_drop_df['feature'].tolist(), axis=1, inplace=True)
print(f"Information Value cut-off being used: {iv_cutoff}\nVariables Dropped: {len(iv_drop_df['feature'].tolist())}. Updated Development Data Shape: {dev_data.shape}")

# Export Feature Dropout Reason
iv_drop_df['drop_reason'] = 'low_iv'
drop_reason_df = pd.read_excel(f'{outpath}/{feature_reduction_outfile}', sheet_name='Drop Reason Summary')
drop_reason_df = drop_reason_df.append(iv_drop_df[['feature', 'drop_reason']])
export_data_to_excel(iv_df, f'{outpath}/{feature_reduction_outfile}', sheetName='Information Value')
export_data_to_excel(drop_reason_df, f'{outpath}/{feature_reduction_outfile}', sheetName='Drop Reason Summary')

<h4> 2.5 Export Coarse Classing result

# Save Updated Development & Validation Data
model_data_pckl_dict['dev_data_out2'] = dev_data
model_data_pckl_dict['val_data_out2'] = val_data[dev_data.columns.tolist()]
with open(f'{outpath}/model_data.pickle', 'wb') as f:
    pickle.dump(model_data_pckl_dict, f)
    
# Coarse Classing Output
coarse_classing_df = coarse_classing_df[coarse_classing_df['VAR_NAME'].isin(dev_data.columns.tolist())]
export_coarse_classing_data(coarse_classing_df, f"{outpath}/{dev_fname.split('.')[0]}_coarse_classing.xlsx", 'coarse_classing_result')
